==> 00_scorecard_defaultrate.Rmd <==
---
title: "Method: Predictors for Institution Graduate Loan Default Rate"
output: pdf_document
---

```{r, echo=FALSE, results=FALSE, warning=FALSE, message=FALSE}
library("knitr")
library(gridExtra)
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, fig.width=5, fig.height=3)
```

```{r results=FALSE}
setwd("/Users/kaylinwalker/R/college_scorecard")
source("/Users/kaylinwalker/R/college_scorecard/02_model.R")
```

### Background  
  
**What makes a certain college a good investment?** Is it producing graduates with less debt? Is it producing graduates who make more money 10 years after finishing? Or is it something less quantifiable, like guiding students into discovering their vocation or completing meaningful coursework?  

The **U.S. Department of Education** quantifies college performance on a number of dimensions. **[College Scorecard](https://collegescorecard.ed.gov)** is an interactive tool designed to help students choose a college based on government data about cost, financial aid, retention, completion and programs. The [data](https://collegescorecard.ed.gov/data/), spanning 1996-2013, is publicly available in .csv format.  


### Summary of Findings    
A linear regression model fit to 4 years of observations (2010-2013; n=6080; training set 80%) found higher completion rate predicted lower default rates, while more Pell Grants predicted higher default rates.    

Graduating students on time (or close) should create graduates more likely to keep their momentum going in finding a job and paying down loans.  

More students with Pell Grants should predict higher rates of loan default since they indicate that students are taking a bigger financial risk (see this [report](http://www.finaid.org/educators/ProfileofPellGrantRecipients.pdf) on financial demographics of Pell Grant recipients; 67% were at or below 150% of the poverty line compared to 16.5% of non-recipients). 

    
### Data & Method  

The model was trained on 4 years of data, 2010-2013, with 20% of the data held out as a validation set. Only four-year public and private non-profit colleges and universities with at least 500 students were included in these figures (1666 schools). 

**Dependent Variable: Default Rate**  
While the dataset includes more than 1700 variables, I decided to focus on the **default rate (3 year cohort)** as the best measure of a college's value. It combines debt with post-graduate earnings into one feature, since a student that defaults on their student loans is clearly either buried in debt, not making high enough salary, or both. The median graduate debt dimension in the dataset wasn't viable since it only includes federal loan debt (which caps at $27,000) and the post-graduate earnings dimension was sparsely populated and therefore not ideal.   

Default rate had a right skewed distribution, which I remedied by doing a log transform. See item 1 in the appendix for a comparison of distributions.     

**Independent Variable Selection**  
I cherry-picked about 30 variables out of the dataset that I thought would have the greatest effect on default rate, including factors related to retention, completion, student debt, financial aid and cost. A large percentage of the variables in the dataset were too specific to have much weight - things like "percent of dependent students who died within 3 years at original institution" and "percent of high-income (above $75,000 in nominal family income) students withdrawn from original institution within 8 years." A lot of variables were missing 90-100% of values and thus had to be thrown out (e.g., "percent of students who submitted a FAFSA to at least one college"). A few variables were excluded after they were found to be highly correlated with other variables (e.g. average family income and Pell Grants; completion and retention).  
  
```{r results=TRUE}
ivs <- data.frame(Variable=names(data3[ c(3:11)]), 
                  Description=c("Public or Private Non-Profit", "Admission Rate", "Size of Enrollment", "Average Cost", "Institutional Expense per Student", "Average Faculty Pay",  "Percentage of Students with Pell Grants", "Completion Rate (6 years)",  "Percentage of Students with Federal Loans"), 
                  Min=c("n/a", min(data3[,4]), min(data3[,5]), min(data3[,6]), min(data3[,7]), min(data3[,8]), min(data3[,9]), min(data3[,10]), min(data3[,11])), 
                  Max= c("n/a", max(data3[,4]), max(data3[,5]), max(data3[,6]), max(data3[,7]), max(data3[,8]), max(data3[,9]), max(data3[,10]), max(data3[,11])), 
                  Units=c("n/a", "0-100", "in thousands", "in thousands", "in thousands", "in thousands", "0-100", "0-100", "0-100"))

kable(ivs)
```

**Missing Value Imputation**  
I only selected variables with less than 25% missing values. Because the Default.2Y and Defaulty.3Y variables were missing values in alternate observations (rarely both missing in an obs.), I was able to combine the 2 year and 3 year default rates and impute the rest. I then used the **mice** package (Multivariate Imputations by Chained Equations) to impute the remaining missing values using the pmm or predictive mean matching method. See item 2 in the appendix for a breakdown of missing values by variable.      
    
### Model      
    
**Model Selection**  
I fit a model with all variables, which found all but institutional expense to be significant, providing an adjusted R-squared value of 0.59, F = 778.6, p < 2.2e-16. Stepwise selection recommended keeping all variables, but keeping only percentage of students with Pell Grants and completion rate resulted in a loss of only 0.015 from adjusted R-squared (Adj R-sq = 0.575, F = 3296). I decided to select this model for simplicity in interpretation. See item 3 in the appendix for a summary of the model; item 4 for the diagnostic plots and item 5 for the predictons vs. actual observations of the test set.   

**Diagnostics**  
The diagnostic plots are mostly normal save for the presence schools with default rates of 0 (29 total, a good chunk of them seminaries or bible schools, but also included California Institute of Technology, Claremont McKenna College, Harvey Mudd College and Vassar College).  
  
**Validation**  
Comparing the model's predicted values to the actual values in the test set produced an R-squared value of 60.50%.   
  
**Coefficient Interpretation**      
Since the dependent variable was log transformed, the coefficients can be interpreted in a slightly different way. Instead of the unit increase in the DV per one unit increase in each variable, the exponentiated (exp) coefficient represents the percent change in the DV per unit chnage in each variable. Those values are shown below, where log.estimate is the coefficient, exp.estimate is exp(coef) and exp.estimate.minus1 is exp(coef) -1, or the percent change.   
  
The factor predicting a lower default rate for a school is completion rate (0.0170% decrease in default rate per one percent increase in completion rate).  
  
The factor predicting higher default rates is Pell Grant percentage (0.0113% increase in default rate per one percent increase in Pell Grants).  
  
  
```{r}
kable(results, row.names=FALSE)
```
  

### Appendix    

1. Default rate transformation  
  
```{r fig.width=8, fig.height=2.5}
a <- ggplot(data3, aes(Default.3Y)) + geom_histogram(stat="bin", binwidth=1) + theme_classic() + labs(title="Default Rate")
b <- ggplot(data3, aes(log(Default.3Y))) + geom_histogram(stat="bin", binwidth=.15) + theme_classic() + labs(title="Log of Default Rate")

grid.arrange(a,b,ncol=2)
```

2. Missing values  
  
```{r}
miss <- data.frame(Variable=c("Admission Rate", "Completion Rate", "Median Debt", 
                               "Institutional Expense per Student","Avg. Faculty Pay", 
                              "Default Rate", "Percent of Students with Pell Grants"), 
                   Imputed=c(529,165,
                             25,31,31,8,4))
kable(miss)
```

3. Summary of linear model  
  
```{r}
summary(fitSmall)
```

4. Linear model diagnostics  
  
```{r fig.height=4, fig.width=6}
par(mfrow=c(2,2))
par(mar=c(2,2,2,2))
plot(fitSmall)
```

5. Linear model validation  
  
```{r}
ggplot(test, aes(log(Default.3Y), predictions)) + geom_point()  + theme_classic() + ylim(c(0,4)) + xlim(c(0,4)) +
    xlab("Log(Default rate)") + ylab("Predicted Value") + labs(title="Predicted Value vs. Actual Value, test set") + 
    geom_abline(slope=1, intercept=0, color="blue") + annotate("text", hjust=1, x=5, y=1, label=paste("R-squared:", round(test.rsq, 4)))
```
